requirements.txt
streamlit
langchain-core
langchain-google-genai
python-dotenv
Pillow
